Alpha status
logging
bug fixes

+++ stitch_images.py
@@ -27,6 +27,7 @@ import sys
 from PIL import Image, ImageDraw, ImageFont
 from PIL.ExifTags import TAGS
 import errno
+import time
 
 ##################################################
 # Settings
@@ -43,36 +44,61 @@ alist_filter = ['JPG','jpg']
 # Output frame rate (frames per second)
 frame_rate = 2
 
+# Log file path
+log_file_path = "."
+
+# Log file name
+log_file_name = "stitch_images_%s.log" % time.strftime("%Y-%m-%d")
+
+# Verbose mode
+verbose = True
+
 ##################################################
+
+# Print to screen (if in verbose mode) and write to the log file
+def update_status(s):
+	if verbose:
+		print s
+	log.write("%s %s\n" % (time.strftime("%H:%M:%S"), s))
+
+
 
 # Setup PIL (Python image library)
 font = ImageFont.truetype("Arial.ttf", 75)
-
-
 
 # Set up command line arguments
 parser = argparse.ArgumentParser()
 parser.add_argument("path", help="top level file path that contains the image folders")
 parser.add_argument("-p", "--prefix", help="image file prefix (default is '%s')" % image_file_prefix)
+parser.add_argument("-l", "--logfilepath", help="log file path (default is '%s')" % log_file_path)
 
 
 args = parser.parse_args()
 rootDir = args.path
 if args.prefix:
 	image_file_prefix = args.prefix
-
-print "\nThe file path is: '%s'\nThe image file prefix is: '%s'" % (rootDir, image_file_prefix)
-
-
+if args.logfilepath:
+	log_file_path = args.logfilepath
+	
+	
 if not os.path.exists(rootDir):
 	print "\nError: The specified file path does not exist.  Exiting the program\n"
 	sys.exit(0)
 
 
+# Make sure path to log file exists before creating the log file
+if not os.path.exists(log_file_path):
+	print "\nError: The specified log file path does not exist.  Exiting the program\n"
+	sys.exit(0)
+log = open(os.path.join(log_file_path, log_file_name), 'a+')
+
+update_status("***** Now starting the image stitch process *****")
+update_status("The file path is: '%s'\nThe image file prefix is: '%s'" % (rootDir, image_file_prefix))
+
 # Start walking through the file tree starting at the specified file path
 for dirName,subdirList,fileList in os.walk( rootDir ) :
   
-	print "Found directory:" , dirName
+	update_status("Found directory: %s" % dirName)
 	#for fname in fileList :
 	#   print "-" , fname
 
@@ -92,7 +118,7 @@ for dirName,subdirList,fileList in os.walk( rootDir ) :
 			# Save the picture paths in an array
 			path_to_file = os.path.join(dirName,file)
 			picture_list[counter].append(path_to_file)	
-			print "  %d) added file %s" % (counter+1, path_to_file)
+			update_status( "  %d) added file %s" % (counter+1, path_to_file))
 			counter = counter + 1
 
 
@@ -102,7 +128,7 @@ for dirName,subdirList,fileList in os.walk( rootDir ) :
 	image_sequence = 0
 	for image_sets in picture_list:
 
-		print "\nNow processing image sequence %d" % (image_sequence+1)
+		update_status( "Now processing image sequence %d" % (image_sequence+1) )
 		
 		# Make sure to remove any pre-existing links to images
 		for i in glob.glob('img*.jpg'):
@@ -110,7 +136,7 @@ for dirName,subdirList,fileList in os.walk( rootDir ) :
 		
 		# Create a directory for the sequence 
 		sequence_dir = os.path.join(dirName,"Seq_%02d" % (image_sequence+1))
-		if not os.path.exists(rootDir):
+		if not os.path.exists(sequence_dir):
 			os.mkdir(sequence_dir)
 		
 		
@@ -127,7 +153,7 @@ for dirName,subdirList,fileList in os.walk( rootDir ) :
 					os.remove(symlink_filepath)
 					os.symlink(image, symlink_filepath)
 					
-			print "     adding image: %s" % image
+			update_status ("     Getting info and creating symlink for image %s" % image)
 			
 			# Get the timestamp from the image
 			# Assume that we don't know it (in case it is unreadable)
@@ -140,10 +166,10 @@ for dirName,subdirList,fileList in os.walk( rootDir ) :
 					timestamp = ctime
 			except: 
 				_type, value, traceback = sys.exc_info()
-				print "Error:\n%r", value
+				update_status( "Error %r" % value )
        						
 			# Add a timestamp image
-			img = Image.new('RGBA',(1200, 100))
+			img = Image.new('RGBA',(900, 100))
 			draw = ImageDraw.Draw(img)
 			draw.text((10, 10), timestamp, font=font, fill=(255, 255, 0))
 			timestamp_filepath = os.path.join(sequence_dir, "timestamp_%010d.png" % counter)
@@ -152,23 +178,33 @@ for dirName,subdirList,fileList in os.walk( rootDir ) :
 			counter = counter+1
 			
 		# Now process a single image sequence
+		# as long there at least one image
+		if counter != 0:
 		
-		#First make a video file of the timestamps
-		ffmpeg_arguments = "ffmpeg -r %s -i %s/timestamp_%%010d.png -sameq -r %s  -vcodec mjpeg  -y %s/_timestamp.mov" % (frame_rate, sequence_dir, frame_rate, sequence_dir)
-		subprocess.call(ffmpeg_arguments, shell=True)    
-		
-		# Then combine the images with the timestamp video
-		#ffmpeg_arguments = "ffmpeg -r 2 -i %s/img%%010d.jpg -sameq -r 2  -vcodec mjpeg   ./seq%03d.mov" % (sequence_dir, image_sequence)
-		ffmpeg_arguments = "ffmpeg -r %s -i %s/img%%010d.jpg -vf \"movie=%s/_timestamp.mov[clip2]; [in][clip2] overlay=0:overlay_h [out]\" -sameq -r %s  -vcodec mjpeg  -y %s/_seq%03d.mov" % (frame_rate, sequence_dir, sequence_dir, frame_rate, sequence_dir, image_sequence+1)
-		subprocess.call(ffmpeg_arguments, shell=True)    
-		
-		# Then remove the timestamp video
-		os.unlink ("%s/_timestamp.mov" % sequence_dir)
+			#First make a video file of the timestamps
+			ffmpeg_arguments = "ffmpeg -r %s -i %s/timestamp_%%010d.png -sameq -r %s  -vcodec mjpeg  -y %s/_timestamp.mov" % (frame_rate, sequence_dir, frame_rate, sequence_dir)
+			update_status(ffmpeg_arguments)
+			subprocess.call(ffmpeg_arguments, shell=True)    
+			
+			# Then combine the images with the timestamp video
+			#ffmpeg_arguments = "ffmpeg -r 2 -i %s/img%%010d.jpg -sameq -r 2  -vcodec mjpeg   ./seq%03d.mov" % (sequence_dir, image_sequence)
+			ffmpeg_arguments = "ffmpeg -r %s -i %s/img%%010d.jpg -vf \"movie=%s/_timestamp.mov[clip2]; [in][clip2] overlay=0:overlay_h [out]\" -sameq -r %s  -vcodec mjpeg  -y %s/_seq%03d.mov" % (frame_rate, sequence_dir, sequence_dir, frame_rate, sequence_dir, image_sequence+1)
+			update_status(ffmpeg_arguments)
+			subprocess.call(ffmpeg_arguments, shell=True)    
+			
+			# Then remove the timestamp video (if it exists)
+			if os.path.exists("%s/_timestamp.mov" % sequence_dir): 
+				os.unlink ("%s/_timestamp.mov" % sequence_dir)
 		
 		image_sequence = image_sequence + 1
 	
     
-# Make sure all temp links to images are removed
+# Make sure all temp links to images are removed - Vince wants these left in
 #for i in glob.glob('img*.jpg'):
 #	os.unlink (i)
+
+update_status("Complete")
+
+
+
